Pull dbUtil functions into db.js
