+++ server/db.js
@@ -16,21 +16,113 @@
 
 //*******************************************************************
 // other files
-const dbUtil = require('./dbUtil.js');
+const include = require('include')(__dirname);
+const models = include('models');
 
 //*******************************************************************
 
-function saveFile(appId, fileInfo, next){
-	dbUtil.saveFile(appId, fileInfo, function(err) {
-		if (err) {
-			return next (err);
-		}
-		else {
-			return next (null);
-		}
-
+function saveFile(applicationId, uploadFile, callback){
+	models.files.create({
+		application_id: applicationId, 
+		file_type: uploadFile.filetypecode, 
+		file_path: uploadFile.keyname,
+		file_name: uploadFile.filename,
+		file_originalname: uploadFile.originalname,
+		file_ext: uploadFile.ext,
+		file_size: uploadFile.size,
+		file_mimetype: uploadFile.mimetype,
+		file_encoding: uploadFile.encoding
+	})
+	.then(function(file) {
+		return callback(null, file);
+	})
+	.catch(function(err) {
+		return callback(err, null);
 	});
 }
+
+const getFiles = function(applicationId, callback){
+
+	models.files.findAll({
+		where: {application_id: applicationId} 
+	})
+	.then(function(files) {
+		return callback(null, files);
+	})
+	.catch(function(err) {
+		return callback(err, null);
+	});
+};
+
+function getApplication(controlNumber, callback){
+	models.applications.findOne({
+		where: {control_number: controlNumber}
+	}).then(function(appl) {
+		if (appl){
+			return callback(null, appl);	
+		}
+		else {
+
+			// TO BE REMOVED begin -- create appl if not exist
+
+			models.applications.findOne({
+				where: {control_number: '1000000000'} 
+			}).then(function(appl) {
+				if (appl){
+					return callback(null, appl);	
+				}
+				else {
+
+					models.applications.create({
+						control_number: '1000000000',
+						form_name: 'FS-2700-3f',
+						region: '11',
+						forest: '22',
+						district: '33',
+						website: 'testwebsite.org',
+						activity_description: 'test activity_description',
+						location_description: 'test location_description',
+						start_datetime: '2017-05-01T10:00:00Z',
+						end_datetime: '2017-05-05T19:00:00Z',
+						number_participants: 45,
+						individual_is_citizen: true,
+						small_business: true,
+						advertising_url: 'www.advertising.com',
+						advertising_description: 'test advertising_description',
+						client_charges: 'test client_charges',
+						experience_list: 'test experience_list'
+					})
+					.then(function(appl) {
+						return callback(null, appl);	
+					})
+					.catch(function(err) {
+						return callback(err, null);
+					});
+				}
+			})
+			.catch(function(err) {
+				return callback(err, null);
+			});
+			
+			// TO BE REMOVED end -- create appl if not exist
+			//return callback('no record found', null);
+		}
+	}).catch(function (err) {
+		console.error(err);
+		return callback(err, null);
+	});
+}
+
+const saveApplication = function(controlNumber, toStore, callback) {
+	models.applications.create(toStore)
+	.then(function(appl) {
+		return callback(null, appl);
+	})
+	.catch(function(err) {
+		console.error(err);
+		return callback(err, null);
+	});
+};
 
 /** Gets list of fields that are to be stored in DB
  * @param  {Object} schema - Schema to look through to find any fields to store in DB
@@ -108,6 +200,9 @@ function getDataToStoreInDB(schema, body){
 	return output;
 }
 
-module.exports.saveFile = saveFile;
 module.exports.getDataToStoreInDB = getDataToStoreInDB;
 module.exports.getFieldsToStore = getFieldsToStore;
+module.exports.saveFile = saveFile;
+module.exports.getFiles = getFiles;
+module.exports.getApplication = getApplication;
+module.exports.saveApplication = saveApplication;
